+++ pokealarmv4/__init__.py
@@ -1,24 +1,11 @@
-import json
-from quart import Quart, request
+from quart import Quart
+from .webviews import general
 
 
-# Fill in app for development purposes
+# Creates a runnable app
 def create_app(test_config=None):
     # create and configure the app
     app = Quart(__name__)
-
-    @app.route('/', methods=['GET'])
-    async def hello():
-        return "PokeAlarm v4 is running!"
-
-    @app.route('/', methods=['POST'])
-    async def receive_events():
-        out = ""
-        data = json.loads(await request.data)
-        ct = 0
-        for event in data:
-            out += f'{ct}: {event}\n'
-            ct += 1
-        return out
+    app.register_blueprint(general)
 
     return app
