+++ pokealarmv4/webviews/__init__.py
@@ -1,0 +1,10 @@
+"""
+pokealarmv4.webviews
+~~~~~~~~~~~~~~~~
+
+This module contains the blueprint for controlling the applications api.
+"""
+
+from .general import general
+
+__ALL__ = [general]
