+++ requirements.txt
@@ -9,4 +9,4 @@ six==1.11.0; python_version < '3'
 typing==3.6.4; python_version < '3'
 
 # v4 Requirements
-quart==0.5.0; python_version >= '3.7'
+quart==0.6.11; python_version >= '3.7'
